+++ firmware/calculations.h
@@ -1,5 +1,5 @@
-extern int expo (int input, unsigned char percent);
-extern void mix(int proportional, int differential, int* a, int* b);
-extern void slowdown (int input, int* output, int increment);
-extern int scale (int input, unsigned int percent);
-extern void limit (int *output, int positive, int negative);
+extern short expo (short input, unsigned char percent);
+extern void mix(short proportional, short differential, short* a, short* b);
+extern void slowdown (short input, short* output, short increment);
+extern short scale (short input, unsigned short percent);
+extern void limit (short *output, short positive, short negative);
